bin/database/requirements.txt
python-dotenv
mysql-connector-python
GitPython
regex